requirements.txt
pandas
numpy
matplotlib
seaborn
tensorflow
scikit-learn
python-dotenv
jupyter